requirements.txt
fastapi
uvicorn
httpx
pyserial
